ignore decode errors when writing data to stdout
In some cases, the serial port may give you garbage data
(such as when the uart changes line speed during bootup).
This raises a python decoding error like this:
  UnicodeDecodeError: 'utf-8' codec can't decode byte 0xa0 in position 0: invalid start byte

Change decoding from the default ('strict') to 'ignore', to
avoid this problem.  See the note inside the program for possible
issues with this approach.