Add command mode
Add the -C option, to use grabserial in "command mode".  This mode
suppresses output that matches the command sent to the serial port
(using the -c argument), and the quit pattern received from the
serial port (using the -q argument).  This has the effect of
producing the output from executing that single command on the
board (if all the data is correct)

This proves to be very useful for running commands against the
NuttX shell (and allowing the option of timing command output, etc.)