Add output file rotation feature (-R)
Allow user to specify a time interval in seconds, to rotate the
output file.  This is done while holding the serial port open,
to avoid data loss.  The previous method of achieving this,
using '-e 3600 -a' closed and reopened the serial port.

Also, clean up a few issues reported by flake8, and move the
version number to 2.1.