Add split-lines (-z) feature
Allow log rotation to split lines of input.  Without this flag,
grabserial will only rotate logs after a newline.  This is useful
for high-speed data.  If data is not coming in at a consistent high
speed, and super-precise log rotation is needed, then use -z or
--split-lines.

I'm not sure this feature is really needed.  Log rotation itself is
a bit of a niche feature.