Add log rotation rounding feature
If units are specified in the log rotation time (s, m, or h),
then the log rotation time will be rounded to a multiple of that
time.  For example, to have the log rotate every half hour, at
the top and bottom of the hour (exactly), use '-R 30m'.

Move the version number to 2.1.2