Add 'putserial' as a work-in-progress
putserial is a new tool to support sending data to a serial port.
The intention is that it is similar to 'cat file >/dev/<ser-dev>'
or 'echo foo >/dev/<ser-dev>' but with the ability to program the
serial port attributes, and handle errors better.

This is a work-in-progress commit, where most of putserial has been
copied from grabserial and not yet refined.