flush outputfd instead of sys.stdout
If grabserial is used as a module and grab() is called with
an argument for outputfd other than sys.stdout, we would be
flushing the wrong file descriptor.  Probably nothing bad
would happen, but it would be the wrong behaviour.

Fix it by using the correct file descriptor to flush in
the main loop.

Signed-off-by: Tim Bird <tim.bird@sony.com>