Handle filepaths correctly for rotated logs
Previously I changed the output filename handling (when -o is used)
to fix a bug with putting the file in the correct directory.  However,
I failed to propogate the needed changes to handle this when the
output filename is rotated.  Make the variable naming more consistent,
by introducing log_filepath, and add appropriate filepath construction
in the log rotation cases (and associated output statements).

Signed-off-by: Tim Bird <tim.bird@sony.com>