Create directory for output log, if needed
Write a new routine that mimics the behavior of 'mkdir -p', and use
that whenever we are about to open up a new logfile.  Note that if
there are time or date specifiers in the directory name portion of
the output log string, then a different directory may be needed
than when the program started (due to log rotation).

Add a test for dynamic directory creation in test-output-filenames.sh

Signed-off-by: Tim Bird <tim.bird@sony.com>