Add option to use different EOL in command string
Add an option to send a different end of line sequence other than
"\n" (LF) in the command string. This is useful to send commands to
devices that parse different EOL in their console, such as
"\r\n" (CRLF) or just a single "\r" (CR). LF is kept as the default.

Refactored a bit by Tim Bird.

Contributed-by: mattmart3 on github
Signed-off-by: Tim Bird <tim.bird@sony.com>